[func] create function
